Term Planner App for WGU C196
A. Create an Android (version 4.0 or higher) mobile application with the following functional requirements:

1. Include the following information for each  term:
•  the term title (e.g., Term 1, Term 2, Spring Term)
•  the start date
•  the end date

2. Include features that allow the user to add as many terms as needed.

3. Implement validation so that a term cannot be deleted if courses are assigned to it.

4. Include features that allow the user to do the following for each  term:
a. add as many courses as needed
b. display a list of courses associated with each  term

c. display a detailed view of each  term, including the following information:
•  the term title (e.g., Term 1, Term 2, Spring Term)
•  the start date
•  the end date

5. Include the following details for each  course:
•  the course title
•  the start date
•  the anticipated end date
•  the status (in progress, completed, dropped, plan to take)
•  the course mentor names, phone numbers, and e-mail addresses

6. Include features that allow the user to do the following for each  course:
a. add as many assessments as needed
b. add a minimum of one optional note per course
c. enter, edit, and delete course information
d. display optional notes
e. display a detailed view of the course, including the due date
f. set alerts for the start and end date, that will trigger when the application is not running
g. share notes via a sharing feature (either e-mail or SMS)

7. Include features that allow the user to do the following for each  assessment:
a. add performance and objective assessments for each  course, including the titles and due dates of the assessments
b. enter, edit, and delete assessment information
c. set alerts for goal dates, that will trigger when the application is not running

B. Design the following screen layouts, including appropriate GUI (graphical user interface) elements (e.g., navigation, input, and information) for each  layout:
•  home screen
•  list of terms
•  list of courses
•  list of assessments
•  detailed course view
•  detailed term view
•  detailed assessment view

C. Create a scheduler and progress tracking application as part of your mobile application from part A.

Note: This can be the implementation of part A.

1. Include the following implementation requirements in the application from part C. Be sure your application is a minimum version 4.0 mobile application:
•  an ArrayList
•  an intent
•  navigation capability between multiple screens using activity
•  three activities
•  events (e.g., a click event)
•  the ability to work in portrait and landscape layout
•  interactive capability (e.g., the ability to accept and act upon user input)
•  a database to store and retrieve application data
•  an application title and an icon
•  notifications or alerts
•  the use of both declarative and programmatic methods to create a user interface

2. Include the following interface requirements in the application from part C:
•  the ability to scroll vertically
•  an action bar
•  two layouts
•  input controls
•  buttons